Added superscript function to all pages
